requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.1
azure-functions==1.12.0
azure-functions-durable==1.2.2
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.14.1
azure-storage-file-datalake==12.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
et-xmlfile==1.1.0
frozenlist==1.3.3
furl==2.1.3
idna==3.4
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.1.0
orderedmultidict==1.0.1
pandas==1.5.2
pipdeptree==2.3.3
portalocker==2.6.0
pyarrow==10.0.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
